Implement nproc_multiplex_tests
This function calls catch_discover_tests on all Catch test cases in the
target, marks them for execution with MPIEXEC, and creates a CTest test
for each "[nproc]" tag in the test name. This tells MPI to execute the
test case with the given number of processes, allowing us to easily
configure each individual test case to be run in a different MPI
enviroment.

Any test case not marked with an "[nproc]" tag is run with all nprocs
possible.

Also, marked all Uniproc tests with "[nproc:1]"